fix: 지원 승인, 거부 실시간 리렌더링 적용
